+++ core/src/de/bsautermeister/snegg/services/DummyGameServices.java
@@ -50,7 +50,7 @@ public class DummyGameServices implements GameServices {
     }
 
     @Override
-    public void submitScore(String leaderboardKey, int highScore) {
+    public void submitScore(String leaderboardKey, long highScore) {
         // NOP
     }
